Update README.md
Title: Predicting California Housing Prices with Linear Regression, Ridge, and Lasso

Description:
The "fetch_california_housing" dataset is a widely used dataset in the field of machine learning and data science. It is a real-world dataset that contains information about housing prices in various districts across California, making it a valuable resource for exploring and applying regression techniques to predict housing prices. This dataset is commonly employed for educational purposes, research, and practical applications in the realm of predictive modeling.